cryptography/safenote/app/requirements.txt
pycrypto
flask
